requirements.txt
# IG Trading Bot - Python Dependencies

# Core
requests>=2.28.0
python-dotenv>=0.20.0

# Data Analysis
pandas>=1.5.0
numpy>=1.23.0

# Financial Data
yfinance>=0.2.0

# PDF Generation
fpdf2>=2.7.0

# Visualization (optional, for charts)
matplotlib>=3.6.0

# Testing
pytest>=7.0.0
